refactor: 선택자 상수 구조 변경
